requirements.txt
boto3==1.35.4
botocore==1.35.4
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
gevent==24.2.1
greenlet==3.0.3
idna==3.7
jmespath==1.0.1
kafka-python==2.0.2
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.0.1
